tests/wpt/web-platform-tests/video-raf/video-request-animation-frame.html

<!DOCTYPE html>
<html>
<title>Test the basics of the video.requestAnimationFrame() API.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<script>
var testVideo = {
  url: getVideoURI('/media/movie_5'),
  height: 240,
  width: 320,
}

async_test(function(t) {
    let video = document.createElement('video');

    let id = video.requestAnimationFrame(
      t.step_func_done((time, metadata) => {
        assert_true(time > 0);
        assert_equals(metadata.height, testVideo.height);
        assert_equals(metadata.width, testVideo.width);
      })
    );

    assert_true(id > 0);

    video.src = testVideo.url;
    video.play();

}, 'Test we can register a video.rAF callback.');


async_test(function(t) {
    let video = document.createElement('video');

    let id = video.requestAnimationFrame(
      t.step_func(_ => {
        assert_unreached("Cancelled callbacks shouldn't be executed")
      })
    );

    video.cancelAnimationFrame(id);

    video.requestAnimationFrame(
      t.step_func(_ => {
        // At this point, the other callback shouldn't have fired, but
        // give it some more time and really make sure it doesn't, by going
        // throught the event loop once more.
        t.step_timeout(() => { t.done(); });
      })
    );

    video.src = testVideo.url;
    video.play();
}, 'Test we can cancel a video.rAF request.');

test(function(t) {
    let video = document.createElement('video');

    // requestAnimationFrame() expects 1 function as a parameter.
    assert_throws(new TypeError(), _ => { video.requestAnimationFrame() } );
    assert_throws(new TypeError(), _ => { video.requestAnimationFrame(0) });
    assert_throws(new TypeError(), _ => { video.requestAnimationFrame("foo") });

    // cancelAnimationFrame() expects 1 number as a parameter
    assert_throws(new TypeError(), _ => { video.cancelAnimationFrame() } );

    // Invalid calls are just noops
    video.cancelAnimationFrame(_ => {});
    video.cancelAnimationFrame(NaN);
    video.cancelAnimationFrame("foo");
    video.cancelAnimationFrame(12345);
    video.cancelAnimationFrame(-1);

}, 'Test invalid calls to the video.rAF API.');
</script>
</html>
